(hg) Report whale-fails
changeset:   2:12b11d15f3f9
user:        John H. Robinson, IV <jaqque@sbih.org>
date:        Mon Oct 04 19:04:34 2010 +0000
files:       extras/Twitter.pl
description:
When twitter whale fails, the json output is no longer json, but
is instead a XHTML with the Overcapacity message.
We look for the URI for the whale fail image to determine if that is the case.

I use that, as it is n9t very likely to be in a legit twit.